feat(cli): modularize CLI and update tests
