+++ crystallize/cli/__init__.py
@@ -1,0 +1,4 @@
+"""CLI package for crystallize."""
+from .app import run
+
+__all__ = ["run"]
